+++ inst/tinytest/test_bootstrap_ca.R
@@ -22,13 +22,13 @@ if (at_home() && Sys.info()["sysname"] == "Linux") {
 
   plot_boot_row <- function() {
     viz_rows(Y, axes = c(1, 2), color = NULL)
-    viz_hull(Y, margin = 1)
+    viz_hull(Y, margin = 1, color = NULL)
   }
   expect_snapshot_plot(plot_boot_row, "CA_boot_row")
 
   plot_boot_col <- function() {
     viz_columns(Y, axes = c(1, 2), color = NULL)
-    viz_hull(Y, margin = 2)
+    viz_hull(Y, margin = 2, color = NULL)
   }
   expect_snapshot_plot(plot_boot_col, "CA_boot_col")
 }